# Conflicts:
#	include/seqan/align/dp_algorithm_impl.h
#	include/seqan/align/dp_align_simd_helper.h
#	include/seqan/align/dp_scout_simd.h
#	include/seqan/align/dp_traceback_impl.h

+++ include/seqan/simd/simd_base_seqan_impl_avx2.h
@@ -147,7 +147,8 @@ _fillVector(TSimdVector & vector,
                 )
             );
 #else
-    vector = SEQAN_VECTOR_CAST_(TSimdVector, _mm256_set_epi64x(std::get<sizeof...(INDICES) - 1 - INDICES>(args)...));
+    vector = SEQAN_VECTOR_CAST_(TSimdVector, 
+                                _mm256_set_epi64x(static_cast<int64_t>(std::get<sizeof...(INDICES) - 1 - INDICES>(args))...));
 #endif
 }
